# Conflicts:
#	assert/complex_assert.cpp
#	assert/double_assert.cpp
#	assert/enum_assert.cpp
#	climate/climate.cpp
#	climate/csv_reader.cpp
#	climate/weather.cpp
#	commercial/hvac.cpp
#	commercial/office.cpp
#	connection/connection.cpp
#	connection/native.cpp
#	connection/udp.cpp
#	connection/udp.h
#	gldcore/convert.cpp
#	gldcore/debug.cpp
#	gldcore/exec.cpp
#	gldcore/find.cpp
#	gldcore/gridlabd.h
#	gldcore/gui.cpp
#	gldcore/legal.cpp
#	gldcore/link/engine/engine.cpp
#	gldcore/load.cpp
#	gldcore/loadshape.cpp
#	gldcore/property.h
#	gldcore/version.h
#	market/collect.cpp
#	market/passive_controller.cpp
#	market/supervisory_control.cpp
#	powerflow/impedance_dump.cpp
#	powerflow/load_tracker.h
#	powerflow/substation.cpp
#	powerflow/transformer.cpp
#	powerflow/transformer_configuration.cpp
#	powerflow/underground_line.cpp
#	residential/house_e.h
#	tape/file.cpp
#	tape/group_recorder.cpp
#	tape/histogram.cpp
#	tape/multi_recorder.cpp
#	tape/player.c
#	tape/recorder.cpp
#	tape/shaper.cpp
#	tape/violation_recorder.cpp
#	tape_plot/tape_plot.cpp

+++ tape/file.cpp
@@ -284,7 +284,7 @@ int file_open_recorder(struct recorder *my, char *fname, char *flags)
 #endif
 	fprintf(my->fp,"# target.... %s %d\n", obj->parent->oclass->name, obj->parent->id);
 	fprintf(my->fp,"# trigger... %s\n", my->trigger[0]=='\0'?"(none)":my->trigger.get_string());
-	fprintf(my->fp, "# interval.. %d\n", static_cast<int>(my->interval));
+	fprintf(my->fp, "# interval.. %lld\n", my->interval);
 	fprintf(my->fp,"# limit..... %d\n", my->limit);
 	fprintf(my->fp,"# timestamp,%s\n", my->property.get_string());
 
@@ -338,7 +338,7 @@ int file_open_collector(struct collector *my, char *fname, char *flags)
 #endif
 	count += fprintf(my->fp,"# group..... %s\n", my->group.get_string());
 	count += fprintf(my->fp,"# trigger... %s\n", my->trigger[0]=='\0'?"(none)":my->trigger.get_string());
-	count += fprintf(my->fp, "# interval.. %d\n", static_cast<int>(my->interval));
+	count += fprintf(my->fp, "# interval.. %lld\n", my->interval);
 	count += fprintf(my->fp,"# limit..... %d\n", my->limit);
 	count += fprintf(my->fp,"# property.. timestamp,%s\n", my->property.get_string());
